requirements.txt
# Web Framework
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Automation
playwright>=1.52.0
selenium>=4.15.0

# AI & ML
ollama>=0.1.7
transformers>=4.35.0
torch>=2.1.0
accelerate>=0.24.0
spacy>=3.7.0

# Document Processing
PyMuPDF>=1.23.0
python-docx>=0.8.11
Pillow>=10.1.0
pdfplumber>=0.9.0

# Data Storage
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Integrations
requests>=2.31.0
httpx>=0.25.0
slack-sdk>=3.23.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0

# NLP
nltk>=3.8.1
textblob>=0.17.1

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
schedule>=1.2.0
asyncio>=3.4.3
aiofiles>=23.2.1
PyYAML>=6.0.1
python-multipart>=0.0.6

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0

# Monitoring
prometheus-client>=0.18.0
structlog>=23.2.0

# Job Application Automation
playwright>=1.40.0
python-dotenv>=0.19.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dateutil>=2.8.2